+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 _Copyright (c) 2014 Daniele Veneroni._  
 _Released under MIT License. See [LICENSE.md](LICENSE.md) for further information._
 
+### Update 31/01/2014 (Version 0.11.0)
+
+* remove element via .remove()
+* Fixed .toggleFullscreen() and expanded browser compatibility
+* get element attribute via .attr('attribute')
+* set element attribute via .attr('attribute', 'value')
+* remove element attribute via .removeAttr('attribute')
+* get the value of a form element via .val()
+* set the value of a form element via .val(value)
+
+* Minor improvements and bugfixs
+
 ### Update 26/01/2014 (Version 0.10.0)
 
 * Tagging version 0.10.0